Develop: Add DB tray, item update on output items

+++ model/Tray.go
@@ -129,7 +129,7 @@ func UpdateTray(trayId int64, trayUpdateRequest TrayUpdateRequest) (int64, error
 	return affected, nil
 }
 
-func UpdateTrayEmpty(trayId int, trayUpdateRequest TrayUpdateRequest) (int64, error) {
+func UpdateTrayEmpty(trayId int64, trayUpdateRequest TrayUpdateRequest) (int64, error) {
 	tx, err := db.BeginTx(context.Background(), nil)
 	if err != nil {
 		return 0, err
